requirements.txt
fastapi
ipympl==0.9.3
ipywidgets==8.1.1
ipywidgets>=8
joblib
jupyter
jupyterlab==4.0.9
matplotlib
pandas
pydantic
scikit-learn
uvicorn
japanize-matplotlib
flask
google-genai
Pillow
python-dotenv
requests
